eventRelated.m: can handle two formats for event timing. 1: the original 'traces' format (myscreen.traces and myscreen.stimtrace contain the temporal information) 2: a simplified 'eventtimes' format (mylog.stimtimes_s is a cell array of the times of each event catgory wrt the onset of the scan)
viewGet.m: when loading the stimulus files determines which of the two formats is applicable. If the stimulus file contains a myscreen variable then it assumes that the file is using 'traces' format. Otherwise it looks for a mylog variable and uses the 'eventtimes' format. The format can be determined by checking the filetype field.