eventRelated.m: handles two formats for event timings. 1: the original 'traces' format (myscreen.traces and myscreen.stimtrace contain the temporal information) 2: a simplified 'eventtimes' format (mylog.stimtimes_s is a cell array of the times of each event catgory wrt the onset of the scan)
viewGet.m: when loading the stimulus files, "viewGet.m" determines which of the two formats (traces vs. eventtimes) is applicable. If the stimulus file contains a non-empty myscreen structure then viewGet returns it and assumes that the file is using the 'traces' format. Otherwise, viewGet return the content of mylog variable and assumes the 'eventtimes' format. The format can be determined by checking the filetype field.