Fixing up Talairach merge.
The exportTal2session callback was not correctly linked from the GUI (somehow the default function anme had been changed)... fixed.
Also, in setTalXform_Callback (in mrAlignGUI), the setting of the matrix "points" which is used to calcualte the vol2tal xform was not getting set properly -- i.e. it had to be modified to properly concatenate the tal points and then take the transpose of the matrix:

  points = [talInfo.AC;talInfo.PC;talInfo.SAC;talInfo.IAC;talInfo.PPC;talInfo.AAC;talInfo.LAC;talInfo.RAC]';

Finally, ALIGN.vol.base no longer contains the full path to the volume anatomy (it shouldn't -- and this was fixed in isbase). So, to open up the correct volume, the field ALIGN.volumePath is now being used.