Fixing 3d pss slice order. There was a bit of confusion here. Before the code had:
procpar.pss = -procpar.pss
and had the rotation matrix basically undoing this.

But that isn't right. fid2nifti always orders slices from smallest to largest pss, so we don't have to reverse order in fid2xform. In fact, we should just assume that is the case so the offset to the first slice is always the min(pss). Which is what the code now does.