This changeset implements plugin functionality for the MLR interface. There are two new added functions:
mlrAdjustGUI, which is the one that plugins call to make adjustments to the GUI (e.g. add menu items)
mlrPlugin, which is used to select/deselect plugins that you want and also installs plugins that are selected

viewGet/viewSet have two new parameters, one to set defaultInterrogators and one to set colormaps. This allows editOverlayGUImrParams to get additional colormaps of your choosing, and allows mrInterrogator to display customized interrogators.

mrOpenWindow has one line in it which calls mlrPlugin to install plugins when you start mrLoadRet.

A default plugin which demonstrates how to write a plugin is in the directory Plugin/Default. It adds a menu item 'Select Plugins'.

mrGetPref was modified to allow two new preferences. One keeps what plugins are selected as a cell array of strings for the folder name (selectedPlugins). The other keeps a comma-delimited list of path names where additional plugin directories may be found.