mlrVol now clips the brightest / darkest points in an image (by default set to the top/bottom 0.2% of pixels, but can be adjusted in the GUI). This helps for display of images that have a stray bright pixel. The readout value when you mouse over points will still display the unclipped value.
getfidk gives a better warning for old files that did not have epi processing run

fid2xform has a few fixes for old data. For one, the epi is not flipped relative to the 2D/3D images, so we don't need to do that. Also, the X dimension is reversed in the old data, so needed to compensate for that