combineTransformOverlays.m: deals with case where scan space and base space are identical but the base is a flat/surface (was r2937 on branch julien)
Conversion of the input overlays to a different space needs to happen if the base spaces are different or if the base is a flat map or surface
(even if the map or surface is from the same base space as the overlay, which can happen if the overlay data is a whole-head structural scan). The latter was not implemented.