refreshMLRDisplay.m, viewGet.m (case ROICache and ROICacheID): fixed surface ROI display for 3D view
In 3D view, any superimposed surface (in mlrAnatomy Plugin) is converted to the current base space. This was not done for surface ROIs.
In addition, the ROI cache was fetched in the current base space (instead of the surface base space), whereas the surface coordinates stay in the base space until display, which also resulted in the wrong ROi coordinates if the ROI was already cached.
I added an optional baseNum parameter to viewGet ROICache and ROICacheID and applied the base transform to the ROI.