mrOpenWindow.m: set slice orientation of current base from what was saved in mrLastView
When opening mrLoadRet, the sliceorientation saved in mrLastView was not
used, so that it was the sliceorientation saved in the base itself that
was used. The base sliceorientation is saved only when switching bases.
If the slice orientation in the view has changed and mrLoadRet is
closed, the last current slice orientation is only saved in the view, not in
the base itself, so it is the view sliceorientation that needs to be
used here