viewSet (case curbase): removed a number of superfluous steps in changing the current base
When changing the current base, a number of steps were redundant:
- setting the rotate and curcoords parameters in the base for the old
  base (effectively,
  the value of the field was obtained using viewGet and then the field
set to the same value using viewSet) These steps used to make send when thes settings were
properties of the view, but not anymore since they have become
properties of a specific base and are set in the base everytime their
value is changed.
- setting the current slice (coords) and the number of slices for the
  new base. This is alsready done when setting the slice orientation in
the view, so doesn't need to be done again.