viewGet.m and viewSet.m: merged curslice and basecurslice cases because the curslice.slicenumfield in the view has become unncessary
The information in view.curslice.slicenum is identical to that stored
in view.baseVolume.curcoords(basesliceindex), so curslice.slicenum has
become redundant. viewGet('curslice') and viewSet('curslice') now
manipulate the base field instead of the view field and are therefore
merged with the already existing basecurslice case. the outdated view
field will be removed in a future commit