mrLoadRetGUI.m: corrected bug with setting the current slice outside the range from the slice editbox
if the current slice was set to something outside the
range using the slice edit box, the slice was corrected to be within the
range in the slice slider (through a call to mlrGuiSet('slicetext')), but not in
the editbox, the display or in the view. Now uses the corrected value from the slider
to set value in editbox, view and display.
Also, when the current slice is changed using the slider, not need to re-read the
value from the slider.