refreshMLRDisplay.m, viewGet.m, viewSet.m: fixed bug with roi cache due to different rotate values used for single and multi-axis.
In multi-axis view, base volume slices are computed for rotate=0, whereas in single-axis view, the rotate value can differ from 0. When an ROI was created in multi-axis view, its coordinates (in base slice space) were always to cache under a name which included the rotate value corresponding to the single-axis view. Therefore an ROI computed and saved to cache in multi-axis view would be wrong when retrieved from the cache in single-axis view (if rotate was differnet from 0) and vice-versa.
I fixed this by propagating the correct rotate value from the base computation to the ROI computation in refreshMLRDisplay.m. This required adding an optional rotate value in calls to viewSet('roicache') and viewSet('roicache').
I also fixed a similar problem with the base-specific information saved in the ROI cache when the same ROI is diplayed on several bases in multibase mode. For this I added an optional basename value in call to viewSet('roicache') and viewGet('roicache').