Changeset allowing to choose between 'Classic' overlay range behaviour and 'New' behaviour implemented in previous commit (a85ac1c)
mrGetPref.m, mrEditPrefs.m: added ovelayRangeBehaviour preference ('Classic' or 'New')
computeOverlay.m: whether the range of the color scale depends on 'range' or 'colorRange' now depends on chosen behaviour
ediOverlayGUI.m: in the 'Classic' behaviour only the clip and range can be set and the colorRange is set to the same values as 'range'. In the 'New' behaviour, the range is called slidersRange and can be set inpendently of colorRange
isoverlay.m, editOverlayGUI: does not require clipping range to be included in the 'range' in  'Classic' mode.
viewGet.m: made the overlay cache ID depend on both the range and colorRange fields because either might change depending on the selected behaviour