Replaced editOverlayGUImrParams.m by the improved version that was located in GLM_v2 Plugin folder (editOverlayGUI.m) and moved the colormaps folder from GLM_v2 for mrLoadRet folder.
This is to avoid having two very similar functions doing the same thing depending on whether the GLM_v2 plugin is selected. editOverlayGUI.m had a slightly different internal organization but basically behaved the same as editOverlayGUImrParams.m, minus a few bugs.
One of the differences is that the color maps are stored in separate files in a colormap folder, which is now mrLoadRet/colormapFunctions (instead of mrLoadRet/Plugin/GLM_v2/colormapFunctions).