Changeset fixing the interrogator when ROIs are visible on a surface (which apparently has been broken since 375c2583, at least in matlab version > 8.4)
In order to fix it, all surface ROI handles are now saved in
view.baseVolumes.hROI (in refreshMLRDisplay). When mrInterrogaotr
initializes, the mrInterrogatorSurfaeCallback is associated not only
with the base surface but with all the ROI surfaces that are displayed,
so that it is called when the user clicks on any of these surfaces (the
bug was due to the fact that the base surface is hidden by the ROI
surfaces). Then in the callback, the current click coordinates are saved
in the base surface and can be retrieved later by mlrGetMouseCoords.