Change set allowing mrLoadRet to be run wihtout a GUI, but still using the last-saved view (mrLastView.mat or any other)
mrLoadRet: added option 'No GUI', which calls mrOpenWindow with noGUI set to false. This then skips all steps that modify the GUI (and require a figure handle) but have not influence on the view.
viewSet and refreshMLRDisplay also skip some GUI-related steps if the the view does not include a figure handle.